Always inject INotificationApiInvoker for separately deployment
