+++ server/alerting/evaluator/src/main/java/org/bithon/server/alerting/evaluator/evaluator/INotificationApiInvoker.java
@@ -16,11 +16,98 @@
 
 package org.bithon.server.alerting.evaluator.evaluator;
 
+import feign.Contract;
+import feign.Feign;
+import feign.codec.Decoder;
+import feign.codec.Encoder;
+import org.bithon.component.commons.concurrency.NamedThreadFactory;
+import org.bithon.component.commons.utils.StringUtils;
+import org.bithon.server.alerting.common.evaluator.EvaluationLogger;
 import org.bithon.server.alerting.notification.api.INotificationApi;
+import org.bithon.server.alerting.notification.message.NotificationMessage;
+import org.bithon.server.discovery.client.DiscoveredServiceInvoker;
+import org.bithon.server.storage.alerting.IEvaluationLogStorage;
+import org.bithon.server.storage.alerting.IEvaluationLogWriter;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.cloud.openfeign.FeignClientsConfiguration;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.core.env.Environment;
+
+import java.util.concurrent.SynchronousQueue;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 /**
  * @author frank.chen021@outlook.com
  * @date 2025/1/5 17:25
  */
 public interface INotificationApiInvoker extends INotificationApi {
+
+    /**
+     * Always inject the notification api invoker 'because it does not strongly rely on the notification service but just a RPC stub
+     */
+    @Configuration
+    @EnableFeignClients
+    @Import(FeignClientsConfiguration.class)
+    class AutoConfiguration {
+
+        @Bean
+        public static INotificationApiInvoker createNotificationAsyncInvoker(ApplicationContext context,
+                                                                             IEvaluationLogStorage logStorage) {
+            INotificationApi impl;
+
+            // The notification service is configured by auto-discovery
+            String service = context.getBean(Environment.class).getProperty("bithon.alerting.evaluator.notification-service", "discovery");
+            if ("discovery".equalsIgnoreCase(service)) {
+                // Even the notification module is deployed with the evaluator module together,
+                // we still call the notification module via HTTP instead of direct API method calls in process
+                // So that it simulates the 'remote call' via discovered service
+                DiscoveredServiceInvoker invoker = context.getBean(DiscoveredServiceInvoker.class);
+                impl = invoker.createUnicastApi(INotificationApi.class);
+            } else if (service.startsWith("http:") || service.startsWith("https:")) {
+                // The service is configured as a remote service at fixed address
+                // Create a feign client to call it
+                impl = Feign.builder()
+                            .contract(context.getBean(Contract.class))
+                            .encoder(context.getBean(Encoder.class))
+                            .decoder(context.getBean(Decoder.class))
+                            .target(INotificationApi.class, service);
+            } else {
+                throw new RuntimeException(StringUtils.format("Invalid notification property configured. Only 'discovery' or URL is allowed, but got [%s]", service));
+            }
+
+            return new INotificationApiInvoker() {
+                // Cached thread pool
+                private final ThreadPoolExecutor notificationThreadPool = new ThreadPoolExecutor(1,
+                                                                                                 10,
+                                                                                                 3,
+                                                                                                 TimeUnit.MINUTES,
+                                                                                                 new SynchronousQueue<>(),
+                                                                                                 NamedThreadFactory.nonDaemonThreadFactory("notification"),
+                                                                                                 new ThreadPoolExecutor.CallerRunsPolicy());
+
+
+                @Override
+                public void notify(String name, NotificationMessage message) {
+                    notificationThreadPool.execute(() -> {
+                        try {
+                            impl.notify(name, message);
+                        } catch (Exception e) {
+                            try (IEvaluationLogWriter writer = logStorage.createWriter()) {
+                                new EvaluationLogger(writer).error(message.getAlertRule().getId(),
+                                                                   message.getAlertRule().getName(),
+                                                                   AlertEvaluator.class,
+                                                                   e,
+                                                                   "Failed to send notification to channel [%s]",
+                                                                   name);
+                            }
+                        }
+                    });
+                }
+            };
+        }
+    }
 }
